bdd: refactor BasicTests into JUnit4
So they can be run via bazel.
